# [vLLM] Support vLLM CPU backend and provide QBits acceleration