--- a/examples/vllm/requirement.txt
+++ b/examples/vllm/requirement.txt
@@ -1,0 +1,3 @@
+accelerate
+datasets
+peft