--- a/intel_extension_for_transformers/transformers/modeling/modeling_auto.py
+++ b/intel_extension_for_transformers/transformers/modeling/modeling_auto.py
@@ -340,16 +340,18 @@ class _BaseQBitsAutoModelClass:
     def from_pretrained(cls, pretrained_model_name_or_path, *model_args, **kwargs):
         use_vllm = kwargs.pop("use_vllm", None)
         if use_vllm is not None:
+            logger.info("The backend is vLLM.")
             from vllm import LLM
             from intel_extension_for_transformers.transformers.utils import RtnConfig
             from intel_extension_for_transformers.transformers.llm.quantization.utils import convert_to_quantized_model
             from vllm.model_executor.model_loader import get_model_loader
+            from vllm.model_executor.model_loader.weight_utils import default_weight_loader
             from vllm.model_executor.layers.linear import (MergedColumnParallelLinear, QKVParallelLinear, RowParallelLinear)
+
             os.environ["backend"] = "use_vllm"
             llm = LLM(model=pretrained_model_name_or_path, trust_remote_code=True)  # Create an vllm instance.
             model = llm.llm_engine.model_executor.driver_worker.model_runner.model
-
-            logger.debug("Original model = ", model)
+            print("Original model =", model)
             class linear_adaptor(torch.nn.Linear):
                 def __init__(self, in_features: int, out_features: int, bias: bool = True, device=None, dtype=None) -> None:
                     super().__init__(in_features, out_features, bias, device, dtype)
@@ -360,8 +362,9 @@ class _BaseQBitsAutoModelClass:
 
             for name, module in model.named_modules():
                 bias_flag = False
-                if isinstance(module, QKVParallelLinear) or isinstance(module, MergedColumnParallelLinear) or isinstance(
-                        module, RowParallelLinear):
+                if isinstance(module, QKVParallelLinear) or isinstance(module, RowParallelLinear):
+                # if isinstance(module, QKVParallelLinear) or isinstance(module, MergedColumnParallelLinear) or isinstance(
+                #         module, RowParallelLinear):
                     out_feature = module.weight.shape[0]
                     in_feature = module.weight.shape[1]
                     if getattr(module, "bias", False) != None:
@@ -380,12 +383,19 @@ class _BaseQBitsAutoModelClass:
 
                     module_traversal._modules[all_module_names[-1]] = copy.deepcopy(torch_linear)
 
-            logger.debug("Optimized model = ", model)
+            print("Optimized model =", model)
             loader = get_model_loader(llm.llm_engine.load_config)
-            model.load_weights(
-                loader._get_weights_iterator(llm.llm_engine.model_config.model,
+
+            weights_iterator = loader._get_weights_iterator(llm.llm_engine.model_config.model,
                                             llm.llm_engine.model_config.revision,
-                                            fall_back_to_pt=True))
+                                            fall_back_to_pt=True)
+
+            # params_dict = dict(model.named_parameters(remove_duplicate=False))
+            # for name in params_dict.keys():
+            #     params = params_dict[name]
+            #     getattr(params, "weight_loader", default_weight_loader)
+
+            model.load_weights(weights_iterator)
 
             config = RtnConfig(compute_dtype="fp32", group_size=128, scale_dtype="fp32", weight_dtype="int4_clip", bits=4)
             model = convert_to_quantized_model(model, config)