## Type of Change

New feature & API change

## Description

- [x] Complete the pipeline to accelerate vLLM linear modules by qbits.
- [x] API Modification: vllm_model = AutoModelForCausalLM.from_pretrained(args.model, use_vllm = True)
- [x] Supported pytorch==2.3.0 + cpu & vLLM==0.4.2 + cpu
- [x] Adapated to more vLLM modules

All models listed in the [vLLM Supported Models](https://docs.vllm.ai/en/latest/models/supported_models.html) can be accelerated theoretically.

Validated models:
* [THUDM/chatglm2-6b](https://hf-mirror.com/THUDM/chatglm2-6b)
* [meta-llama/Llama-2-7b-chat-hf](https://hf-mirror.com/meta-llama/Llama-2-7b-chat-hf)
* [baichuan-inc/Baichuan2-7B-Chat](https://hf-mirror.com/baichuan-inc/Baichuan2-7B-Chat)
* [tiiuae/falcon-7b](https://huggingface.co/tiiuae/falcon-7b)
* [meta-llama/Meta-Llama-3-8B](https://huggingface.co/meta-llama/Meta-Llama-3-8B)
* [mistralai/Mistral-7B-Instruct-v0.1](https://huggingface.co/mistralai/Mistral-7B-Instruct-v0.1)
* [microsoft/phi-2](https://huggingface.co/microsoft/phi-2)
* [Qwen/CodeQwen1.5-7B-Chat](https://huggingface.co/Qwen/CodeQwen1.5-7B-Chat)

Performance:
```text
# Env: SPR numactl -m 0 -C 0-55
# Model: Llama2-7b-chat-hf

32 in, 32 out
The Neural Speed:  942 ms  (Generation Speed: 33.97 toks/s, 213ms first token * 1+ 23.5ms next token * 31)
The qbits    vLLM: 1677 ms (Generation Speed: 19.08 toks/s)
The original vLLM: 2304 ms (Generation Speed: 13.75 toks/s)

1024 in 32 out
The Neural Speed  : 2344ms  (Generation Speed: 13.65 toks/s, 1527ms first token * 1+ 31 next_token * 26.38ms)
The qbits    vLLM : 2912ms  (Generation Speed: 11.0 toks/s)
The original vLLM : 3575ms  (Generation Speed: 8.95 toks/s)
```

## Expected Behavior & Potential Risk

N/A

## How has this PR been tested?

Nightly unit tests.

Llama2-7b
![image](https://github.com/intel/intel-extension-for-transformers/assets/109137058/856d4bad-fd4f-4b3c-8790-03a6866f2c39)

## Dependency Change?

added in the example/vllm/requirement.txt